Add dependency from CAPI to THLO and THLO->gml_st.
It seems that the conversion pass for MHLO->THLO is *in MHLO*, which seems wrong. Since that is transitively coming it, we then need to build/link THLO. Then THLO is depending on interfaces defines in gml_st. This seems like really suspect layering. I am just doing the minimum to make it build.